Added key listeners to all keys again and on keypress action to adjust color
Changed name of child component of all keys containing the text_mesh to
Key_Text.
Added new class KeyActivator which is responsible for listenting for a
specific key event and chage that keys color when the event triggers. Also
restors the orginal color on keyUp. Id to listen for and color to switch
to are variables of the behaviour.
Updated keyboard prefab to reflect latest state.