Can now pick which item to manipulate
When a GameObject is touched by leapmotion hands, it becomes the
selected object. The selected object can be rotated, translated, and
scaled by hand gestures.

To enable an item for selection add the Item Selection script component
and ensure that there is an ItemManipulator component on some object in
the scene.